requirements.txt
django
django-redis
requests
requests-oauthlib
django-cron
djangorestframework
